move txMsg functionality to versionedTransaction
