+++ light-sdk-ts/src/test-utils/sendVersionedTransaction.ts
@@ -1,12 +1,49 @@
-import { VersionedTransaction } from "@solana/web3.js";
+import {
+  AddressLookupTableAccount,
+  ComputeBudgetProgram,
+  TransactionMessage,
+  VersionedTransaction,
+} from "@solana/web3.js";
 import * as anchor from "@coral-xyz/anchor";
 import { Provider } from "wallet";
 import { confirmConfig } from "../constants";
 
-export const sendVersionedTransaction = async (
-  compiledTx: anchor.web3.MessageV0,
-  provider: Provider,
-) => {
+export const sendVersionedTransaction = async (ix: any, provider: Provider) => {
+  const recentBlockhash = (
+    await provider.provider!.connection.getRecentBlockhash("confirmed")
+  ).blockhash;
+
+  const txMsg = new TransactionMessage({
+    payerKey: provider.relayer.accounts.relayerPubkey,
+    instructions: [
+      ComputeBudgetProgram.setComputeUnitLimit({ units: 1_400_000 }),
+      ix,
+    ],
+    recentBlockhash: recentBlockhash,
+  });
+
+  const lookupTableAccount = await provider.provider!.connection.getAccountInfo(
+    provider.relayer.accounts.lookUpTable,
+    "confirmed",
+  );
+
+  const unpackedLookupTableAccount = AddressLookupTableAccount.deserialize(
+    lookupTableAccount!.data,
+  );
+
+  const compiledTx = txMsg.compileToV0Message([
+    {
+      state: unpackedLookupTableAccount,
+      key: provider.relayer.accounts.lookUpTable,
+      isActive: () => {
+        return true;
+      },
+    },
+  ]);
+
+  compiledTx.addressTableLookups[0].accountKey =
+    provider.relayer.accounts.lookUpTable;
+
   var tx = new VersionedTransaction(compiledTx);
   let retries = 3;
   let res;
