+++ light-sdk-ts/src/test-utils/testRelayer.ts
@@ -44,45 +44,7 @@ export class TestRelayer extends Relayer {
   async sendTransaction(instruction: any, provider: Provider): Promise<any> {
     try {
       if (!provider.provider) throw new Error("no provider set");
-
-      const recentBlockhash = (
-        await provider.provider.connection.getRecentBlockhash("confirmed")
-      ).blockhash;
-
-      console.log({ recentBlockhash });
-
-      const txMsg = new TransactionMessage({
-        payerKey: provider.relayer.accounts.relayerPubkey,
-        instructions: [
-          ComputeBudgetProgram.setComputeUnitLimit({ units: 1_400_000 }),
-          instruction,
-        ],
-        recentBlockhash: recentBlockhash,
-      });
-
-      const lookupTableAccount =
-        await provider.provider.connection.getAccountInfo(
-          provider.lookUpTable!,
-          "confirmed",
-        );
-
-      const unpackedLookupTableAccount = AddressLookupTableAccount.deserialize(
-        lookupTableAccount!.data,
-      );
-
-      const compiledTx = txMsg.compileToV0Message([
-        {
-          state: unpackedLookupTableAccount,
-          key: provider.relayer.accounts.lookUpTable!,
-          isActive: () => {
-            return true;
-          },
-        },
-      ]);
-
-      compiledTx.addressTableLookups[0].accountKey = provider.lookUpTable!;
-
-      const response = await sendVersionedTransaction(compiledTx, provider);
+      const response = await sendVersionedTransaction(instruction, provider);
       return response;
     } catch (err) {
       console.error("erorr here =========>", { err });
